+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * [#2800](https://github.com/clojure-emacs/cider/pull/2800): Add support for force-out debugger command.
 * Add support for nREPL 0.8 `completions` op. It's used if `cider-nrepl` is not available.
 * Add `browser` to the list of supported ClojureScript REPL types.
+* Add an interactive command to toggle Clojure font-locking in the REPL (`cider-repl-toggle-clojure-font-lock`).
 
 ### Changes